// src/app/models/language.model.ts
export interface Language {
    id: number;
    name: string;
    icon: string;
    createdBy: string;
    createdAt: string;
    isDelete: boolean;
    languageCode: string;
    updatedBy: string;
    updatedAt: string;
}



/* createdAt: "2022-03-18 04:44:59"
createdBy: "e61a39db-0fc1-4af4-a50d-d18955fd7928"
icon: "759b7d77-e314-4cc7-a230-cc02e4f62eae.png"
id: 1
isDelete: false
languageCode: "lo"
name: "ລາວ"
updatedAt: "2022-03-18 04:44:59"
updatedBy: "e61a39db-0fc1-4af4-a50d-d18955fd7928" */